requirements.txt
torch
torchvision
numpy
pandas
matplotlib
seaborn
scikit-learn
jupyter
pillow
tqdm
setuptools
wheel
pip
GitPython
pydicom

-e .